Change all long int to int32_t
The controller runs on a Raspberry Pi, where sizeof(long int) is 4
bytes. I'm running OpenModelica on a 64bit machine. From now on, let's
always compile the controller with -m32, and forget about this.

+++ SocketLibrary/include/ControlBuffer.h
@@ -2,14 +2,15 @@
 #define __BUFFER_CONTROL_H
 
 #include <GeneralBuffer.h>
+#include <stdint.h>
 
 typedef struct control_buffer *ControlBuffer;
 
 ControlBuffer CB_init(const int size);
 int CB_destroy(ControlBuffer b);
 
-int CB_getControl(ControlBuffer b, long int *res);
-int CB_setControl(ControlBuffer b, const long int * const val);
+int CB_getControl(ControlBuffer b, int32_t *res);
+int CB_setControl(ControlBuffer b, const int32_t * const val);
 
 GBuffer CB_getBuffer(ControlBuffer b);
